requirements.txt
torch
torchvision
timm
pillow
numpy
pandas
scikit-learn
tqdm
matplotlib